+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 CHANGELOG for Sulu Sales
 ========================
+
 * dev-develop
 
+    * FEATURE     [CoreBundle]    Changed visual appearance of addons in item-table.
     * FEATURE     [CoreBundle]    Added option to not display currency in item-table.
     * BUGFIX      [CoreBundle]    Displaying first found tax class in item tables overlay.
     * FEATURE     [CoreBundle]    Display overlay (on click) also when item-table is disabled.